smv int -> double

+++ BonusCalcListener/Boundary/WorkOrderOperativeSmvData.cs
@@ -10,7 +10,7 @@ namespace BonusCalcListener.Boundary
         public string Address { get; set; }
 
         // The SMV for the whole job
-        public int StandardMinuteValue { get; set; }
+        public double StandardMinuteValue { get; set; }
         public string OperativeId { get; set; }
         public double? JobPercentage { get; set; }
         public DateTime? ClosedTime { get; set; }
